# replace Sprintf with FormatInt for performance