fmt.Sprintf is not performing well

name   old time/op  new time/op  delta
New-4   283µs ± 1%   244µs ± 3%  -13.71%  (p=0.000 n=17+18)